Use vertex data directly out of a buffer, if possible.  This requires that the caller configured the vertex arrays in native hardware format; otherwise the vertices need to be converted and copied into the command list as usual.
CVA works as normal with this, so you can lock a set of arrays in
buffers, and use the cached converted vertices.  However, the CVA
state is not associated with the vertex buffers, so you still need to
re-lock the arrays if you switch them around.  This is something which
might be addressed in the future (maintain a cache of locked array
sets, maybe?).